+++ code/ecommerce-loyalty/src/main/java/com/ecommerce/loyalty/query/LoyaltyQueryService.java
@@ -43,5 +43,5 @@ public interface LoyaltyQueryService {
      * @param userId the user ID
      * @return the level multiplier
      */
-    double getMemberMultiplier(Long userId);
+    BigDecimal getMemberMultiplier(Long userId);
 }
